+++ assets/redux/monitoring/monitoringApi.js
@@ -62,6 +62,26 @@ export const monitoringApi = createApi({
         };
       },
     }),
+    getTurnoverByType: builder.query({
+      query: (queryConfig) => {
+        const { type, ...filters } = queryConfig;
+        const queryParams = queryString.stringify(filters, { skipNull: true, arrayFormat: 'bracket' });
+        return {
+          url: `/api/turnover/${type}?${queryParams}`,
+          method: 'GET',
+        };
+      },
+    }),
+    getTurnoverChart: builder.query({
+      query: (queryConfig) => {
+        const { chartResolution, ...filters } = queryConfig;
+        const queryParams = queryString.stringify(filters, { skipNull: true, arrayFormat: 'bracket' });
+        return {
+          url: `/api/turnover/chart/${chartResolution}?${queryParams}`,
+          method: 'GET',
+        };
+      },
+    }),
     getUsersList: builder.query({
       query: (queryConfig) => {
         const queryParams = queryString.stringify(queryConfig, { skipNull: true, arrayFormat: 'bracket' });
@@ -169,6 +189,8 @@ export const {
   useGetEventsLatestQuery,
   useGetEventTypesQuery,
   useGetTurnoverQuery,
+  useGetTurnoverByTypeQuery,
+  useGetTurnoverChartQuery,
   useGetUsersListQuery,
   useConfirmUserMutation,
   useSwitchToAdminMutation,